refactor: barrel exports and updated imports accordingly
