+++ src/components/footer/footer.jsx
@@ -1,4 +1,4 @@
-import "./footer.css";
+import "./Footer.css";
 import { Link } from "react-router-dom";
 export function Footer() {
   return (
