+++ src/contexts/index.js
@@ -1,0 +1,4 @@
+export { NavbarProvider } from "./navbar-context"
+export { ProductProvider } from "./product-context"
+export { useFilters } from "./product-context"
+export { useNavbar } from "./navbar-context"
