+++ src/pages/home/home.css
@@ -19,8 +19,6 @@
   width:100%;
   justify-content: space-around;
   flex-wrap: wrap;
-  /* display: grid;
-  grid-template-columns: repeat(auto-fill, 16.5rem); */
   gap: var(--spacing-sm);
 
 }
