+++ src/pages/home/home.jsx
@@ -1,7 +1,5 @@
 import "./home.css";
-import { CategoryList } from "../../components/CategoryList/CategoryList";
-import { BestDeals } from "../../components/BestDeals/BestDeals";
-import { Banner } from "../../components/Banner/Banner";
+import { Banner, CategoryList, BestDeals } from "../../components/index";
 export function Home() {
   return (
     <main className="page-main">
